% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXLabel.R
\name{SXLabel}
\alias{SXLabel}
\title{SXLabel}
\usage{
SXLabel(inputId, label = NULL, styleclass = "default", link = NULL)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{The label text.}

\item{styleclass}{A Bootstrap style to apply to the label. (\code{default, primary, success, info, warning, or danger}).}

\item{link}{An optional html link associated with the label.}
}
\description{
helper function to include labels in shiny apps
}

