% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXPanel.R
\name{SXPanel}
\alias{SXPanel}
\title{SXPanel}
\usage{
SXPanel(inputId, ..., heading = "", text_size = NULL,
  styleclass = "default", checkbox = FALSE, collapsible = FALSE,
  removeButton = FALSE, icon = NULL)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{...}{Panel contents.}

\item{heading}{Panel heading.}

\item{text_size}{Size for panel heading (NULL = do not change text size, 'small' = 75\% size, 'large' = 150\% size).}

\item{styleclass}{A Bootstrap style to apply to the button. (\code{default, primary, success, info, warning, or danger}).}

\item{checkbox}{\bold{logical} - add checkbox to left of panel header.}

\item{collapsible}{\bold{logical} - Create a collapsible panel.}

\item{removeButton}{\bold{logical} - Include a button to remove the panel.}

\item{icon}{Name of a glyphicon icon to include to the right of panel header.}
}
\description{
helper function to include panels in shiny apps
}

