% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXSortable.R
\name{SXSortable}
\alias{SXSortable}
\title{SXSortable}
\usage{
SXSortable(inputId, labels = NULL, styleclass = "default")
}
\arguments{
\item{inputId}{The input slot that will be used to access the data.}

\item{labels}{A list of labels.}

\item{styleclass}{A Bootstrap style to apply to the list. (\code{default, primary, success, info, warning, or danger}).}
}
\description{
add a set of labels and make them sortable
}

