% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXTextArea.R
\name{SXTextArea}
\alias{SXTextArea}
\title{SXTextArea}
\usage{
SXTextArea(inputId, label = NULL, text = NULL, placeholder = NULL,
  resizable = TRUE, rows = 5, cols = 40)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{text}{Text present in the textarea.}

\item{placeholder}{Placeholder text for the textarea.}

\item{resizable}{\bold{logical} - Allow textarea to be resized (default = true).}

\item{rows}{number of rows (defaults to 5).}

\item{cols}{number of columns (defaults to 40px).  100\% covers the width of the parent.}
}
\description{
helper function to include textarea in shiny apps
}

